requirements.txt
streamlit>=1.35.0
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
plotly>=5.18.0
streamlit-drawable-canvas>=0.9.3
scikit-learn>=1.3.0
